+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Re-publish state every scan to ensure fresh
 - Main discovery and state message now better in sync on `in_progress` flag
 - State not published during update if Home Assistant mode off
+### Version Detection
+- Version policy can be also be set in `custom_packages.yaml` or in the `packages` section of config
 ## 1.8.0
 ### Package Info
 - The `common_packages.yaml` that ships inside the Docker image can be extended, or overridden, using the `packages` section of the main config file